// src/app/model/tarjetaEspecialidad.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TarjetaEspecialidad { 
    idTarjetaEspcialidad?: number;
    tipoEspecialidad?: TarjetaEspecialidad.TipoEspecialidadEnum;
}
export namespace TarjetaEspecialidad {
    export type TipoEspecialidadEnum = 'ESPECIALIADADES_GASTRO_CIRUJIA_GENERAL' | 'GINECOLOGIA' | 'MEDICINA_GENERAL' | 'PEDIATRIA' | 'PSICOLOGIA';
    export const TipoEspecialidadEnum = {
        ESPECIALIADADESGASTROCIRUJIAGENERAL: 'ESPECIALIADADES_GASTRO_CIRUJIA_GENERAL' as TipoEspecialidadEnum,
        GINECOLOGIA: 'GINECOLOGIA' as TipoEspecialidadEnum,
        MEDICINAGENERAL: 'MEDICINA_GENERAL' as TipoEspecialidadEnum,
        PEDIATRIA: 'PEDIATRIA' as TipoEspecialidadEnum,
        PSICOLOGIA: 'PSICOLOGIA' as TipoEspecialidadEnum
    };
}